+++ example/provider.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 
@@ -5,11 +7,25 @@ void main() {
   runApp(MyApp());
 }
 
+class Bloc {
+  final StreamController<int> _streamController = StreamController();
+  Stream<int> stream;
+
+  Bloc() {
+    stream = _streamController.stream.asBroadcastStream();
+  }
+
+  void dispose() {
+    _streamController.close();
+  }
+}
+
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Provider.value(
-      value: 42,
+    return Provider<Bloc>(
+      builder: (_) => Bloc(),
+      dispose: (_, value) => value.dispose(),
       child: Example(),
     );
   }
@@ -18,6 +34,11 @@ class MyApp extends StatelessWidget {
 class Example extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return Text(Provider.of<int>(context).toString());
+    return StreamBuilder<int>(
+      stream: Provider.of<Bloc>(context).stream,
+      builder: (context, snapshot) {
+        return Text(snapshot.data?.toString() ?? 'Foo');
+      },
+    );
   }
 }
