+++ example/provider_value.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/widgets.dart';
+import 'package:provider/provider.dart';
+
+void main() {
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Provider.value(
+      value: 42,
+      child: Example(),
+    );
+  }
+}
+
+class Example extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Text(Provider.of<int>(context).toString());
+  }
+}
